fix JS behaviour
