+++ pages/parts/carusel.html
@@ -2,7 +2,7 @@
   <!--       <a class="next">&lt</a>-->
   <button id="prevBtn" class="btn btn-left">&lt;&lt;</button>
   <div class="aspect-ratio-box">
-    <img class="carousel__sliderImg" />
+    <img class="carousel__sliderImg" alt="Сейчас вылетит птичка" />
   </div>
   <button id="nextBtn" class="btn btn-right">&gt;&gt;</button>
 </div>
